Rework tests to unify development and testing
Summary:
The previous tests used beaker to create a VM and run the manifests, which is
silly, since quite separately there is a Vagrantfile to create a development
environment in which to run the manifests.

Now, the tests use the same Vagrantfile to create a VM, then run the tests in
them, and make assertions with serverspec directly.

This serves as a proof of concept for our new CI strategey. Ref T4.

Test Plan: Tests run on my workstation. Just "rake spec" does the job. One fails (testing idempotency), but this is an issue that already existed, but wasn't tested.

Reviewers: #macpros_internal

Maniphest Tasks: T4

Differential Revision: http://phabricator.macprofessionals.lan/D5

+++ Vagrantfile
@@ -1,24 +1,17 @@
 Vagrant.configure("2") do |config|
-  config.vm.box = "ubuntu-server-12042-x64-vbox4210"
-  config.vm.box_url = "http://puppet-vagrant-boxes.puppetlabs.com/ubuntu-server-12042-x64-vbox4210.box"
+  config.vm.box = "ubuntu-server-1310-x64-virtualbox"
+  config.vm.box_url = "http://puppet-vagrant-boxes.puppetlabs.com/ubuntu-1310-x64-virtualbox-puppet.box"
 
   config.vm.network :forwarded_port, guest: 80, host: 8080
 
-  config.vm.provider "virtualbox" do |v|
-    v.customize ["modifyvm", :id, "--memory", "1024"]
-  end
-
   config.vm.provision "shell", inline: <<-EOS
     apt-get -qq update
-    apt-get install -y git
-    /opt/ruby/bin/gem install --no-ri --no-rdoc librarian-puppet
+    apt-get install -y git ruby-dev build-essential libxml2-dev libxslt1-dev
+    gem install --no-ri --no-rdoc librarian-puppet bundler
     mkdir -p /usr/share/puppet/modules
-    ln -s /vagrant /usr/share/puppet/modules/phabricator
-    (cd /usr/share/puppet/modules/phabricator && /opt/ruby/bin/librarian-puppet install --path /usr/share/puppet/modules)
+    [ ! -e /usr/share/puppet/modules/phabricator ] && ln -s /vagrant /usr/share/puppet/modules/phabricator
+    (cd /usr/share/puppet/modules/phabricator && librarian-puppet install --path /usr/share/puppet/modules)
+    (cd /vagrant && bundle)
+    touch /etc/puppet/hiera.yaml
   EOS
-
-  config.vm.provision :puppet do |puppet|
-    puppet.manifests_path = "vagrant"
-    puppet.manifest_file = "init.pp"
-  end
 end
